+++ netqasm/lang/instr/instr_enum.py
@@ -77,6 +77,9 @@ class Instruction(Enum):
     RET_REG = 39
     RET_ARR = 40
 
+    # Move source qubit to target qubit (target is overwritten)
+    MOV = 41
+
 
 _COMMAND_GROUPS = {
     RegCommand: [